refactor: Extract SceneTimeseriesMenu
